Added the Swal alert to delete and Update functionality

+++ client/src/components/List.jsx
@@ -1,10 +1,15 @@
 import React, { useEffect, useState } from "react";
 import { MdDelete } from "react-icons/md";
 import { FaEdit } from "react-icons/fa";
-const List = ({ displayForm, setUpdateForm, updateForm }) => {
+import Swal from "sweetalert2";
+const List = ({
+  displayForm,
+  setUpdateForm,
+  updateForm,
+  setSelectedTodoId,
+}) => {
   const [list, setList] = useState([]);
   const [error, setError] = useState("");
-
   const getAllList = async () => {
     try {
       const response = await fetch(`http://localhost:3000`);
@@ -24,17 +29,33 @@ const List = ({ displayForm, setUpdateForm, updateForm }) => {
   }, [displayForm, updateForm]);
 
   const deleteList = async (id) => {
-    const response = await fetch(`http://localhost:3000/delete/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
+    const confirmation = await Swal.fire({
+      icon: "warning",
+      title: "Are you sure?",
+      text: "You wont/'t be able to revert this!",
+      showCancelButton: true,
+      confirmButtonColor: "red",
+      cancelButtonColor: "blue",
+      confirmButtonText: "Delete",
+      cancelButtonText: "Cancel",
     });
-    const result = await response.json();
-    if (!response.ok) {
-      setError(result.error);
+    if (confirmation.isConfirmed) {
+      try {
+        const response = await fetch(`http://localhost:3000/delete/${id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        const result = await response.json();
+        if (!response.ok) {
+          setError(result.error);
+        }
+        getAllList();
+      } catch (err) {
+        setError(err);
+      }
     }
-    getAllList();
   };
   return (
     <section className="bg-amber-100 grow">
@@ -56,7 +77,10 @@ const List = ({ displayForm, setUpdateForm, updateForm }) => {
                 <FaEdit
                   size={25}
                   className="cursor-pointer"
-                  onClick={() => setUpdateForm(true)}
+                  onClick={() => {
+                    setUpdateForm(true);
+                    setSelectedTodoId(item._id);
+                  }}
                 />
                 <MdDelete
                   size={25}
